Refatoração de código

+++ database/migrations/2019_02_27_210253_create_family_members_table.php
@@ -48,7 +48,7 @@ class CreateFamilyMembersTable extends Migration
                 'stepsister'
             ]);
             $table->string('work');
-            $table->double('income');
+            $table->decimal('income', 10, 2);
             $table->integer('assisted_id')->unsigned();
             $table->foreign('assisted_id')->references('id')->on('assisteds');
             $table->softDeletes();
